+++ python/neuroglancer/token.py
@@ -1,0 +1,5 @@
+import os
+import binascii
+
+def make_random_token():
+  return binascii.hexlify(os.urandom(20))
