+++ python/neuroglancer/volume.py
@@ -1,0 +1,29 @@
+# @license
+# Copyright 2016 Google Inc.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from token import make_random_token
+
+class ServedVolume(object):
+  def __init__(self, data, offset=(0,0,0), channel_type=None, voxel_size=(1,1,1)):
+    self.token = make_random_token()
+    self.data = data
+    self.voxel_size = voxel_size
+    self.offset = offset
+    self.shape = data.shape[::-1]
+    self.data_type = data.dtype.name
+    if self.data_type == 'uint8':
+      self.channel_type = 'image'
+    else:
+      self.channel_type = 'annotation'
+
